Перемещает функции для рендера в отдельный файл render.js. Функция рендер теперь принимает на вход компонент, а не элемент
