+++ src/utils.js
@@ -1,8 +1,3 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-};
-
 const castTimeFormat = (value) => {
   return value < 10 ? `0${value}` : String(value);
 };
@@ -30,22 +25,5 @@ const getMarkupFromArray = (array, elementCreator) => {
     .join(`\n`);
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
 
-  return newElement.firstChild;
-};
-
-const render = (container, element, position) => {
-  switch (position) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export {castTimeFormat, getFormatTime, getFormatDate, getRandomBoolean, getMarkupFromArray, createElement, render, RenderPosition};
+export {castTimeFormat, getFormatTime, getFormatDate, getRandomBoolean, getMarkupFromArray};
